moved the hydra generation out to a util file and started to do the same with the audio, haven't quite completed that. added meytda to the program as well.
