+++ client/package.json
@@ -11,6 +11,7 @@
   },
   "dependencies": {
     "hydra-synth": "^1.3.29",
+    "meyda": "^5.6.2",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
     "react-router-dom": "^6.20.1"
